+++ embassy-stm32/CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - feat: stm32/usart: add `eager_reads` option to control if buffered readers return as soon as possible or after more data is available ([#4668](https://github.com/embassy-rs/embassy/pull/4668))
 - feat: stm32/usart: add `de_assertion_time` and `de_deassertion_time` config options
 - change: stm32/uart: BufferedUartRx now returns all available bytes from the internal buffer
+- change: stm32/(q|o|x)spi: Set the default memory size to the maximum allowable for qspi/ospi/xspi so that the indirect read command is able to read more than 2 bytes. This avoids a potential gotcha.
 
 ## 0.4.0 - 2025-08-26